refactor: split confluence.ts into directory with one file per tool
Split the monolithic 1296-line confluence.ts file into a well-organized
directory structure with one file per tool for better maintainability:

Structure:
- packages/tools/src/web/confluence/
  - index.ts (main exports and factory function)
  - types.ts (ConfluenceToolsConfig interface)
  - auth.ts (authentication helper functions)
  - search-confluence.ts (searchConfluence tool)
  - get-confluence-page.ts (getConfluencePage tool)
  - list-confluence-spaces.ts (listConfluenceSpaces tool)
  - get-space-pages.ts (getSpacePages tool)
  - create-confluence-page.ts (createConfluencePage tool)
  - update-confluence-page.ts (updateConfluencePage tool)
  - archive-confluence-page.ts (archiveConfluencePage tool)

Benefits:
- Each file has a single, clear responsibility
- Easy to find and modify individual tools
- Better for code review
- Easier to test individual tools
- Follows single responsibility principle

All 7 tools are now in separate files (~100 lines each) instead of
one massive 1296-line file. Build passes successfully.

Part of Confluence tools migration - Phase 4
